requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.4
requests==2.31.0
openpyxl==3.1.2
Pillow==10.2.0